requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

attrs>=18.2.0,<=19.3.0
click~=7.0
docker~=3.7
frozendict~=1.2
gitpython~=2.1
murmurhash3~=2.3.4
packaging>=14.0,<=19.2
pre-commit>=1.0.0,<=1.18.3
psutil~=5.6.3
pyyaml>=5.1.2
semantic-version~=2.8.0
tqdm~=4.36.1
validate-email~=1.3
